refactor: 책임 분리 및 패키지 구조 재정의
