+++ src/main/java/com/programmers/java/calculator/CalculatorApplication.java
@@ -1,23 +1,21 @@
 package com.programmers.java.calculator;
 
-import com.programmers.java.calculator.converter.PostfixExpressionConverter;
-import com.programmers.java.calculator.io.Console;
-import com.programmers.java.calculator.model.Calculator;
+import com.programmers.java.calculator.calculator.PostfixExpressionCalculator;
+import com.programmers.java.calculator.controller.CalculatorController;
 import com.programmers.java.calculator.repository.MemoryCalculatorRepository;
 import com.programmers.java.calculator.service.CalculatorService;
-import com.programmers.java.calculator.service.CalculatorServiceImpl;
-import com.programmers.java.calculator.validation.ExpressionValidator;
-import com.programmers.java.calculator.validation.Validator;
+import com.programmers.java.calculator.util.PostfixExpressionConverter;
+import com.programmers.java.calculator.view.Console;
 
 public class CalculatorApplication {
 
     public static void main(String[] args) {
         Console console = new Console();
-        CalculatorService calculatorService = new CalculatorServiceImpl(
+        CalculatorService calculatorService = new CalculatorService(
                 new PostfixExpressionConverter(),
+                new PostfixExpressionCalculator(),
                 new MemoryCalculatorRepository());
-        Validator validator = new ExpressionValidator();
 
-        new Calculator(console, console, calculatorService, validator).run();
+        new CalculatorController(console, console, calculatorService).run();
     }
 }
