+++ src/main/java/com/programmers/java/calculator/calculator/Calculator.java
@@ -1,0 +1,6 @@
+package com.programmers.java.calculator.calculator;
+
+public interface Calculator<T, U> {
+
+    U calculate(T t);
+}
