+++ src/main/java/com/programmers/java/calculator/docs/README.md
@@ -1,0 +1,26 @@
+# 요구사항 분석
+
+## Domain
+
+## view
+
+### Console
+
+- 입력과 출력을 담당하는 기능
+- 추후 GUI로 변경 가능성이 있기 때문에 Interface 구현
+
+```markdown
+입력
+
+- 메뉴(조회, 연산, 종료) 선택 입력
+- 수식 입력
+
+출력
+
+- 에러 메시지 출력
+- 계산 결과 출력
+- 저장된 계산 이력 출력
+```
+
+## Controller
+
