+++ src/main/java/com/programmers/java/calculator/domain/Operator.java
@@ -1,4 +1,4 @@
-package com.programmers.java.calculator.model;
+package com.programmers.java.calculator.domain;
 
 import java.math.BigDecimal;
 import java.util.Collections;
