--- src/main/java/com/programmers/java/calculator/validation/Validator.java
@@ -1,6 +1,0 @@
-package com.programmers.java.calculator.validation;
-
-public interface Validator {
-
-    void validate(Object object);
-}
