+++ src/test/java/com/programmers/java/calculator/MenuSelectionTest.java
@@ -1,6 +1,6 @@
 package com.programmers.java.calculator;
 
-import com.programmers.java.calculator.model.MenuType;
+import com.programmers.java.calculator.domain.MenuType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
